feat. 1.16.5移植第一次提交

+++ src/main/java/top/gregtao/concerto/config/ClientConfig.java
@@ -1,8 +1,8 @@
 package top.gregtao.concerto.config;
 
 import com.google.gson.GsonBuilder;
-import net.minecraft.client.util.math.Vector2f;
 import top.gregtao.concerto.enums.TextAlignment;
+import top.gregtao.concerto.util.Vector2i;
 
 public class ClientConfig extends ConfigFile {
     public static ClientConfig INSTANCE = new ClientConfig();
@@ -79,8 +79,8 @@ public class ClientConfig extends ConfigFile {
             return this.y.getPosition(height);
         }
 
-        public Vector2f getPos(int width, int height) {
-            return new Vector2f(this.getX(width), this.getY(height));
+        public Vector2i getPos(int width, int height) {
+            return new Vector2i(this.getX(width), this.getY(height));
         }
     }
